refactor: shared components
