+++ web/src/lib/components/shared-components/settings/setting-checkboxes.svelte
