+++ web/src/lib/components/shared-components/settings/setting-switch.svelte
